Rename class and fix compilation warnings.
