+++ app/src/main/java/emg/soccerstats/api/FootbalDataService.kt
@@ -1,11 +1,12 @@
-package emg.soccerstats.interfaces
+package emg.soccerstats.api
 
+import emg.soccerstats.models.Competition
 import emg.soccerstats.models.CompetitionsModel
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface RetrofitService {
+interface FootbalDataService {
   @GET("/v2/competitions")
   fun competitionsService(
     @Query("X-Auth-Token") authToken: String
